Added events handling, CPU in same mode as GPU, runtime config via config.ini file

+++ ElectricFieldGPU/ElectricFieldImageCreator.cpp
@@ -23,15 +23,15 @@ void ElectricFieldImageCreator::UpdateImage(float * electricFieldMatrix)
 			if (level <= 0)
 				image.setPixel(i, j, sf::Color(0, 0, 0));
 			else if (level <= 255)
-				image.setPixel(i, j, sf::Color(0, 0, level));
+				image.setPixel(i, j, sf::Color(0, 0, static_cast<int>(level)));
 			else if (level <= 510)
-				image.setPixel(i, j, sf::Color(0, level - 255, 255));
+				image.setPixel(i, j, sf::Color(0, static_cast<int>(level - 255), 255));
 			else if (level <= 765)
-				image.setPixel(i, j, sf::Color(level - 510, 255, 765 - level));
+				image.setPixel(i, j, sf::Color(static_cast<int>(level - 510), 255, static_cast<int>(765 - level)));
 			else if (level <= 1020)
-				image.setPixel(i, j, sf::Color(255, 1020 - level, 0));
+				image.setPixel(i, j, sf::Color(255, static_cast<int>(1020 - level), 0));
 			else if (level <= 1149)
-				image.setPixel(i, j, sf::Color(1276 - level, 0, 0));
+				image.setPixel(i, j, sf::Color(static_cast<int>(1276 - level), 0, 0));
 			else
 				image.setPixel(i, j, sf::Color(127, 0, 0));
 		}
